requirements.txt
httpx[http2]>=0.24.0
cloudscraper
beautifulsoup4>=4.12.0
openai
python-dotenv
brotli